requirements.txt
boto3==1.10.45
botocore==1.13.45
docutils==0.15.2
jmespath==0.9.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.7
